[TASK]Security Bug Fix

+++ Classes/Controller/BackupsController.php
@@ -1,10 +1,11 @@
 <?php
 namespace NITSAN\NsBackup\Controller;
 
-use TYPO3\CMS\Core\Core\Environment;
-use TYPO3\CMS\Extbase\Object\ObjectManager;
 use NITSAN\NsBackup\Domain\Repository\BackupglobalRepository;
+use RuntimeException;
+use TYPO3\CMS\Core\Core\Environment;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Extbase\Object\ObjectManager;
 use TYPO3\CMS\Extbase\Utility\LocalizationUtility as transalte;
 
 /***
@@ -28,7 +29,8 @@ class BackupsController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControll
      */
     protected $backupglobalRepository;
     protected $backupBaseController = null;
-    protected $errorValidation = null;
+    protected $errorValidation      = null;
+    protected $objectManager        = null;
 
     /**
      * Inject the BackupglobalRepository repository
@@ -52,13 +54,13 @@ class BackupsController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControll
     {
         // Global error check
         //@extensionScannerIgnoreLine
-        $this->objectManager = GeneralUtility::makeInstance(ObjectManager::class);
-        $this->backupBaseController = $this->objectManager->get(BackupBaseController::class);
+        $this->objectManager        = GeneralUtility::makeInstance(ObjectManager::class);
+        $this->backupBaseController = $this->objectManager->get(\NITSAN\NsBackup\Controller\BackupBaseController::class);
 
         $this->errorValidation = $this->backupBaseController->globalErrorValidation();
-        if(!empty($this->errorValidation)) {
-            $header = transalte::translate('global.errorvalidation','ns_backup');
-            $message = transalte::translate('global.errorvalidation.message','ns_backup');
+        if (! empty($this->errorValidation)) {
+            $header  = transalte::translate('global.errorvalidation', 'ns_backup');
+            $message = transalte::translate('global.errorvalidation.message', 'ns_backup');
             $this->addFlashMessage($message, $header, \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
         }
     }
@@ -70,20 +72,25 @@ class BackupsController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControll
      */
     public function dashboardAction()
     {
+        // Load JavaScript modules (equivalent to v11-13 functionality)
+        $pageRenderer = $this->objectManager->get(\TYPO3\CMS\Core\Page\PageRenderer::class);
+        $pageRenderer->loadRequireJsModule('TYPO3/CMS/NsBackup/jquery');
+        $pageRenderer->loadRequireJsModule('TYPO3/CMS/NsBackup/Main');
+
         $globalSettingsData = $this->backupglobalRepository->findAll();
-        $arrBackupData = $this->backupglobalRepository->findBackupDataAll(5);
-        $arrMultipleVars = [
-            'cleanup' => constant('cleanup'),
-            'backuptype' => constant('backuptype'),
-            'compress' => constant('compress'),
-            'backupglobal' => $globalSettingsData[0],
-            'action' => 'dashboard',
-            'arrBackupData' => $arrBackupData,
-            'errorValidation' => $this->errorValidation
+        $arrBackupData      = $this->backupglobalRepository->findBackupDataAll(5);
+        $arrMultipleVars    = [
+            'cleanup'         => constant('cleanup'),
+            'backuptype'      => constant('backuptype'),
+            'compress'        => constant('compress'),
+            'backupglobal'    => ! empty($globalSettingsData[0]) ? $globalSettingsData[0] : null,
+            'action'          => 'dashboard',
+            'arrBackupData'   => $arrBackupData,
+            'errorValidation' => $this->errorValidation,
         ];
-       // @extensionScannerIgnoreLine
+        // @extensionScannerIgnoreLine
         if (version_compare(TYPO3_branch, '11', '>=')) {
-            $arrMultipleVars['modalAttr'] ='data-bs-';
+            $arrMultipleVars['modalAttr'] = 'data-bs-';
         } else {
             $arrMultipleVars['modalAttr'] = 'data-';
         }
@@ -97,42 +104,88 @@ class BackupsController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControll
      */
     public function backuprestoreAction()
     {
+        $pageRenderer = $this->objectManager->get(\TYPO3\CMS\Core\Page\PageRenderer::class);
+        $pageRenderer->loadRequireJsModule('TYPO3/CMS/NsBackup/jquery');
+        $pageRenderer->loadRequireJsModule('TYPO3/CMS/NsBackup/Main');
+
         $globalSettingsData = $this->backupglobalRepository->findAll();
+        if (! empty($globalSettingsData[0])) {
+            $globalBackupStorePath = $globalSettingsData[0]->getBackupStorePath();
+            $isPublicPath          = $this->isPathPublic($globalBackupStorePath);
+        }
 
         $arrMultipleVars = [
-            'cleanup' => constant('cleanup'),
-            'backuptype' => constant('backuptype'),
-            'compress' => constant('compress'),
-            'backupglobal' => $globalSettingsData[0],
-            'action' => 'backuprestore',
-            'errorValidation' => $this->errorValidation
+            'cleanup'         => constant('cleanup'),
+            'backuptype'      => constant('backuptype'),
+            'compress'        => constant('compress'),
+            'backupglobal'    => ! empty($globalSettingsData[0]) ? $globalSettingsData[0] : null,
+            'action'          => 'backuprestore',
+            'errorValidation' => $this->errorValidation,
         ];
 
-        $arrPost = $this->request->getArguments();
+        $arrPost    = $this->request->getArguments();
+        $backupName = trim($arrPost['backuprestore']['backupName'] ?? '');
+        if (! empty($backupName) && preg_match('/[^0-9A-Za-z _-]/', $backupName)) {
+            $sanitizedName = htmlspecialchars($backupName, ENT_QUOTES, 'UTF-8');
+
+            $this->addFlashMessage(
+                "Invalid backup name: '{$sanitizedName}'. " . transalte::translate('manualbackup.error.description', 'ns_backup'),
+                transalte::translate('manualbackup.error', 'ns_backup'),
+                \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR
+            );
+
+            $this->redirect('backuprestore');
+            return;
+        }
 
         // "RUN" Backup from "Manual Backup Module"
-        $arrPost['backuprestore'] = $arrPost['backuprestore'] ?? '';
-        $arrPost = $arrPost['backuprestore'];
-        if(!empty($arrPost['backupFolderSettings']) && empty($this->errorValidation)) {
+        $arrPost = $arrPost['backuprestore'] ?? '';
+
+        if (! empty($arrPost['backupFolderSettings']) && empty($this->errorValidation)) {
 
             // Create json and take backup
-            $arrResponse = $this->backupBaseController->generateBackup($arrPost);
-            if($arrResponse['log'] == 'error') {
+            try {
+                $arrResponse = $this->backupBaseController->generateBackup($arrPost);
+            } catch (RuntimeException $e) {
+                $this->addFlashMessage($e->getMessage(), transalte::translate('manualbackup.error', 'ns_backup'), \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
+                $this->redirect('backuprestore');
+                return;
+            }
+            if ($arrResponse['log'] == 'error') {
                 // Error Flash-Message
-                $mesHeader = transalte::translate('manualbackup.error','ns_backup');
+                $mesHeader   = transalte::translate('manualbackup.error', 'ns_backup');
                 $backup_file = $arrResponse['backup_file'];
                 $this->addFlashMessage($backup_file, $mesHeader, \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
-            }
-            else {
-                // Success Flash-Message
-                $mesHeader = transalte::translate('manualbackup.success','ns_backup');
-                $backup_file = transalte::translate('backup.downloaded','ns_backup').' '.$arrResponse['backup_file'];
+            } else {
+                $mesHeader   = transalte::translate('manualbackup.success', 'ns_backup');
+                $backup_file = transalte::translate('backup.downloaded', 'ns_backup') . ' ' . $arrResponse['backup_file'];
                 $this->addFlashMessage($backup_file, $mesHeader);
-
-                // Pass to Fluid
+                $response = (array) json_decode($arrResponse['log']);
+                if (isset($response['errorCount']) && $response['errorCount'] > 0) {
+                    $globalSettingsData = $this->backupglobalRepository->findAll();
+                    if (! empty($globalSettingsData[0]) && $globalSettingsData[0]->getEmailNotificationOnError()) {
+                        $emails = GeneralUtility::trimExplode(',', $globalSettingsData[0]->getEmails(), true);
+                        foreach ($emails as $email) {
+                            $mail = GeneralUtility::makeInstance(\TYPO3\CMS\Core\Mail\MailMessage::class);
+                            $mail->setFrom([$globalSettingsData[0]->getEmailFrom() => 'Backup']);
+                            $mail->setTo([$email]);
+                            $mail->setSubject($globalSettingsData[0]->getEmailSubject());
+                            $mail->setBody('<p><strong>Backup Error:</strong> \'' . $response['errors'][0]->message . '\'</p>', 'text/html');
+                            $mail->send();
+                        }
+                        $this->addFlashMessage(
+                            $response['errors'][0]->message,
+                            transalte::translate('manualbackup.warning', 'ns_backup'),
+                            \TYPO3\CMS\Core\Messaging\AbstractMessage::WARNING
+                        );
+                    }
+                }
                 $arrMultipleVars['isManualBackup'] = '1';
-                $arrMultipleVars['log'] = '<pre class="pre-scrollable"><code class="json">'. json_encode(json_decode($arrResponse['log']), JSON_PRETTY_PRINT) .'</code></pre>';
-                $arrMultipleVars['download_url'] = $arrResponse['download_url'];
+                $arrMultipleVars['log']            = '<pre class="pre-scrollable"><code class="json">' . json_encode(json_decode($arrResponse['log']), JSON_PRETTY_PRINT) . '</code></pre>';
+                $arrMultipleVars['download_url']   = '';
+                if ($isPublicPath) {
+                    $arrMultipleVars['download_url'] = $arrResponse['download_url'];
+                }
             }
         }
         // List Backup History
@@ -142,17 +195,16 @@ class BackupsController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControll
             if ($objBackupData[$keyBackup]['logs']) {
                 $objBackupData[$keyBackup]['logs'] = '<pre class="pre-scrollable"><code class="json">' . json_encode(json_decode($objBackupData[$keyBackup]['logs']), JSON_PRETTY_PRINT) . '</code></pre>';
             }
-            if($valueBackup['download_url']){
-                $file_headers = @get_headers($valueBackup['download_url']);
-                $objBackupData[$keyBackup]['isDownload'] = (!$file_headers || $file_headers[0] == 'HTTP/1.1 404 Not Found') ? false : true;
+            if ($valueBackup['download_url']) {
+                $file_headers                            = @get_headers($valueBackup['download_url']);
+                $objBackupData[$keyBackup]['isDownload'] = (! $file_headers || $file_headers[0] == 'HTTP/1.1 404 Not Found') ? false : true;
             }
         }
         //@extensionScannerIgnoreLine
         if (version_compare(TYPO3_branch, '11', '>=')) {
-            $arrMultipleVars['modalAttr'] ='data-bs-';
-
+            $arrMultipleVars['modalAttr'] = 'data-bs-';
         } else {
-            $arrMultipleVars['modalAttr'] ='data-';
+            $arrMultipleVars['modalAttr'] = 'data-';
         }
         $arrMultipleVars['arrBackupData'] = $objBackupData;
         $this->view->assignMultiple($arrMultipleVars);
@@ -165,31 +217,56 @@ class BackupsController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControll
     public function deletebackupbackupAction()
     {
         $uid = GeneralUtility::_GP('uid');
+
+        $globalSettingsData = $this->backupglobalRepository->findAll();
+        $globalSettingsData = ! empty($globalSettingsData[0]) ? $globalSettingsData[0] : null;
+
+        if (! $globalSettingsData) {
+            $headerMsg = transalte::translate('something.wrong.here', 'ns_backup');
+            $this->addFlashMessage($headerMsg, '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR, true);
+            return $headerMsg;
+        }
+
         $arrBackup = $this->backupglobalRepository->findBackupByUid($uid);
         // Let's delete it
         $this->backupglobalRepository->removeBackupData($uid);
+
         // Remove file from Physical location
-        if(file_exists($arrBackup['filenames'])){
+        if (file_exists($arrBackup['filenames'])) {
             unlink($arrBackup['filenames']);
         }
 
-        $rootPath = $this->globalSettingsData[0]->root ?? (Environment::getProjectPath() ?? '');
-        if(Environment::isComposerMode()) {
-            $rootPath = Environment::getPublicPath();
+        if (Environment::isComposerMode()) {
+            $rootPath = Environment::getPublicPath() ?: PATH_site;
         }
-        $jsonFolder = $rootPath.'/uploads/tx_nsbackup/json/';
-        if(file_exists($jsonFolder.$arrBackup['jsonfile'])) {
-            unlink($jsonFolder.$arrBackup['jsonfile']);
-        }
-
-        $jsonLogFile=str_replace("_configuration","_log",$arrBackup['jsonfile']);
-        if(file_exists($jsonFolder.$jsonLogFile)) {
-            unlink($jsonFolder.$jsonLogFile);
+        $rootPath   = $globalSettingsData->getBackupStorePath() ?? ($rootPath . '/uploads');
+        $jsonFolder = $rootPath . '/tx_nsbackup/json/';
+        if (file_exists($jsonFolder . $arrBackup['jsonfile'])) {
+            unlink($jsonFolder . $arrBackup['jsonfile']);
         }
 
-        $headerMsg = transalte::translate('delete.backup.data','ns_backup');
-        $msg = transalte::translate('delete.backup.message','ns_backup').$arrBackup['filenames'];
-        $this->addFlashMessage($msg, $headerMsg);
+        $jsonLogFile = str_replace("_configuration", "_log", $arrBackup['jsonfile']);
+        if (file_exists($jsonFolder . $jsonLogFile)) {
+            unlink($jsonFolder . $jsonLogFile);
+        }
+
+        $headerMsg = transalte::translate('delete.backup.data', 'ns_backup');
+        $msg       = transalte::translate('delete.backup.message', 'ns_backup') . $arrBackup['filenames'];
+        $this->addFlashMessage($msg, $headerMsg, \TYPO3\CMS\Core\Messaging\AbstractMessage::OK, true);
         return $msg;
+    }
+
+    /**
+     * @param string $path
+     * @return boolean
+     */
+    public function isPathPublic(string $path): bool
+    {
+        if (! Environment::isComposerMode()) {
+            $valuesToCheck = ['typo3', 'typo3conf', 'vendor', 'typo3temp', 'bin'];
+            $parts         = array_filter(explode('/', rtrim($path, '/')));
+            return empty(array_intersect($parts, $valuesToCheck));
+        }
+        return str_contains(rtrim($path, '/'), Environment::getPublicPath());
     }
 }
